requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
ecdsa==0.19.0
fastapi==0.114.1
gunicorn==23.0.0
h11==0.14.0
idna==3.8
packaging==24.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pymssql==2.3.1
pypyodbc==1.3.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
